Handle non-Google NDK builds when extracting version from RELEASE.TXT
No change of behavior with NDKs that managed to make it through the
CMake process before - this just keeps CMake from erroring out in
computing ANDROID_NDK_RELEASE_NUM on some alternate NDKs. The full
logic follows:

If there is no release file, or if it is parseable as a standard one,
only then do we compute ANDROID_NDK_RELEASE_NUM. Otherwise, if there's
one there but it's not a standard one, we grab the first token of it
(to drop 64-bit), stick that in ANDROID_NDK_RELEASE instead, and set
ANDROID_NDK_RELEASE_THIRD_PARTY. If that's set, a warning is printed
at the end of the file, because some such NDKs (like CrystaX NDK) require
other build changes that are integrated into the makefiles but not this
script yet.